@Zetanova 2021-08-28
The foreach is not good too.
It is creating a try-catch

Because the Transport.Addresses is of type ISet<Address> the simple and obvious `Contains` method is the best fit
```
public bool HasAddress(Address address)
{
         retrun address.Equals(_local.RootPath.Address) || address.Equals(RootPath.Address) 
                    || Transport.Addresses.Contains(address);
}
```

see: https://sharplab.io/#gist:b0a2cc23e7748f71577070b2748e462e