@Aaronontheweb 2021-08-30
Great suggestions @Zetanova - I'll update the PR with some numbers for those later this week.