@Aaronontheweb 2021-08-31
```
OSVersion:                         Microsoft Windows NT 6.2.9200.0
ProcessorCount:                    16
ClockSpeed:                        0 MHZ
Actor Count:                       32
Messages sent/received per client: 200000  (2e5)
Is Server GC:                      True
Thread count:                      111

Num clients, Total [msg], Msgs/sec, Total [ms]
         1,  200000,    106952,    1870.69
         5, 1000000,    197981,    5051.82
        10, 2000000,    198709,   10065.92
        15, 3000000,    197239,   15210.15
        20, 4000000,    196329,   20374.52
        25, 5000000,    196634,   25428.24
        30, 6000000,    195015,   30767.75        
```

Final numbers look good to me!