@Zetanova 2021-08-31
The equals itself is not optimal, the Port nullable can be better compared.

The EqualityComparer<Address>.Default.Equlas(left, right) can be used,
it should have a better perf then a 2x branching with an own code (not sure)

See code at:
https://sharplab.io/#gist:f028291550b1c641019aca408efdbc3a