@Aaronontheweb 2021-09-01
> The equals itself is not optimal, the Port nullable can be better compared.
> 
> The `EqualityComparer<Address>.Default.Equlas(left, right)` can be used,
> it should have a better perf then a 2x branching with an own code (not sure)
> 
> See code at:
> https://sharplab.io/#gist:f028291550b1c641019aca408efdbc3a

```csharp
  public bool Equals(Address other)
    {
        return other != null
            && string.Equals(_host, other._host) 
            && string.Equals(_system, other._system) 
            && string.Equals(_protocol, other._protocol)
            && (_port ?? 0) == (other._port ?? 0);
    }
```

You want the port comparison to go first - as it's the fastest route to exit the loop for two values who are not equal. You also still want to check for equality by reference in this method in case it's called directly via the `.Equals` method