@Aaronontheweb 2021-09-01
Final numbers:

```
OSVersion:                         Microsoft Windows NT 6.2.9200.0
ProcessorCount:                    16
ClockSpeed:                        0 MHZ
Actor Count:                       32
Messages sent/received per client: 200000  (2e5)
Is Server GC:                      True
Thread count:                      112

Num clients, Total [msg], Msgs/sec, Total [ms]
         1,  200000,    108755,    1839.86
         5, 1000000,    200201,    4995.26
        10, 2000000,    199442,   10028.27
        15, 3000000,    198900,   15083.07
        20, 4000000,    199552,   20045.01
        25, 5000000,    198413,   25200.80
        30, 6000000,    195224,   30734.52
```

Looks to me like the technical errors have been resolved as well.