@to11mtm 2021-09-03
> 
> 
> The equals itself is not optimal, the Port nullable can be better compared.
> 
> The `EqualityComparer<Address>.Default.Equlas(left, right)` can be used,
> it should have a better perf then a 2x branching with an own code (not sure)
> 
> See code at:
> https://sharplab.io/#gist:f028291550b1c641019aca408efdbc3a

I'm not sure... the ASM generated looks too sparse for what we're doing here. I'm not sure if Sharplab is acccurate here.

I would expect something like this would be the 'best' way to do ==:

```
    public static bool operator ==(Address left, Address right)
    {
            if (left != null) {
                return left.Equals(right);
            }
            return right == null;
    }
``` 

This is based on how [`.Equals()` itself is implemented in GenericEqualityComparer](https://referencesource.microsoft.com/#mscorlib/system/collections/generic/equalitycomparer.cs,129)