@Zetanova 2021-09-07
I looked at the whole class 

The `RootPath.Address` is the same as `_local.RootPath.Address`
currently the equals will perform twice on the same address
```csharp
public bool HasAddress(Address address)
{
         retrun address.Equals(_local.RootPath.Address) || Transport.Addresses.Contains(address);
}
```

then the cache extension instances do not need to be volatile,
every memory barrier in a hot path should be avoided:
```csharp
private ActorRefResolveThreadLocalCache _actorRefResolveThreadLocalCache;
private ActorPathThreadLocalCache _actorPathThreadLocalCache;

public virtual void Init(ActorSystemImpl system)
{
            _system = system;

            //moved up
            _actorRefResolveThreadLocalCache = ActorRefResolveThreadLocalCache.For(system);
            _actorPathThreadLocalCache = ActorPathThreadLocalCache.For(system);

            _local.Init(system);

            //rest of code
}
``` 

Then `ActorPath.TryParse` is getting called multiple times (with and without cache)

I made an draft at https://github.com/akkadotnet/akka.net/pull/5273
It is still WIP