@Aaronontheweb 2021-09-07
> The RootPath.Address is the same as _local.RootPath.Address currently the equals will perform twice on the same address

Not a huge deal from a perf point of view - `ReferenceEquals` call is about 1-2 ns.