@Zetanova 2021-09-07
@Aaronontheweb  I made some improvements, volatile should/can be removed from ` private Internals _internals;` too.
This makes in a hot-path a difference. 

Then I found a way to remove AddressCache and use the same ActorPathCache as RemoteActorRefProvider